requirements.txt
flask
scikit-learn==1.6.1
pandas
numpy
gunicorn
pytesseract
Pillow
PyMuPDF
python-docx
